requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
ecdsa==0.18.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
passlib==1.7.4
pip==24.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.29.0
